+++ client/.gitignore
@@ -20,4 +20,4 @@
 # yarn-debug.log*
 # yarn-error.log*
 
-client/src/components/*.css
+*.css
