+++ client/package.json
@@ -17,7 +17,7 @@
     "build": "yarn sass-compile && react-scripts build",
     "test": "react-scripts test",
     "eject": "react-scripts eject",
-    "sass-compile": "node-sass src/stylesheets/_components -o src/components/"
+    "sass-compile": "node-sass src/stylesheets/_components -o src/assets/css/"
   },
   "eslintConfig": {
     "extends": "react-app"
