+++ client/src/Dashboard/containers/DashboardContainer.js
@@ -1,5 +1,5 @@
 import {connect} from 'react-redux';
-import Dashboard from '../components/dashboard/Dashboard';
+import Dashboard from '../components/Dashboard';
 
 const mapStateToProps = state => {
   return {};
