+++ client/src/Public/containers/CreateAccountContainer.js
@@ -1,6 +1,6 @@
 import {connect} from 'react-redux';
 import CreateAccount from '../components/CreateAccount';
-import {createUser, createUserReset, userCreationStatus as status} from '../actions/createUser';
+import {createUser, createUserReset, userCreationStatus as status} from '../../actions/createUser';
 
 const mapStateToProps = state => {
   const {userCreationStatus, errorMsg} = state.userCreation;
