+++ client/src/Public/containers/LoginContainer.js
@@ -1,6 +1,6 @@
 import {connect} from 'react-redux';
 import Login from '../components/Login';
-import {loginUser, loginStatus as status} from '../actions/authenticate';
+import {loginUser, loginStatus as status} from '../../actions/authenticate';
 
 const mapStateToProps = state => {
   const {loginStatus, errorMsg} = state.userAuthentication;
