--- client/src/components/App.css
@@ -1,39 +1,0 @@
-header {
-  display: flex;
-  justify-content: center; }
-
-.msg, body {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center; }
-
-.button {
-  display: inline-block;
-  padding: 10px 16px;
-  border: 1px solid #333;
-  text-align: center;
-  cursor: pointer; }
-  .button:hover {
-    background-color: rgba(255, 255, 255, 0.1); }
-
-.msg {
-  color: darkred;
-  padding: 10px 16px; }
-
-html {
-  height: 100%;
-  background-color: snow; }
-
-body {
-  min-height: 100%;
-  position: relative; }
-
-header {
-  position: absolute;
-  top: 0;
-  left: 0;
-  right: 0;
-  padding: 20px 32px;
-  background-color: #777;
-  color: snow; }
