--- client/src/components/Login.css
@@ -1,37 +1,0 @@
-.msg {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center; }
-
-.button {
-  display: inline-block;
-  padding: 10px 16px;
-  border: 1px solid #333;
-  text-align: center;
-  cursor: pointer; }
-  .button:hover {
-    background-color: rgba(255, 255, 255, 0.1); }
-
-.msg {
-  color: darkred;
-  padding: 10px 16px; }
-
-section {
-  padding: 20px 32px;
-  background-color: lightgreen;
-  min-width: 345px;
-  text-align: center; }
-  section div {
-    padding: 10px 16px; }
-
-input {
-  display: block;
-  box-sizing: border-box;
-  padding: 2px;
-  margin: 10px 0;
-  width: 100%; }
-
-.button {
-  margin: 0 4px;
-  min-width: 44px; }
