+++ client/src/index.js
@@ -4,7 +4,7 @@ import thunk from 'redux-thunk';
 import {Provider} from 'react-redux';
 import {createStore, applyMiddleware} from 'redux';
 import rootReducer from './reducers';
-import App from './components/App';
+import App from './App';
 
 const store = createStore(rootReducer, applyMiddleware(thunk));
